requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
colorama==0.4.5
Django==4.0
django-auth-ldap==4.1.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-mssql-backend==2.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==14.2.0
flake8==5.0.4
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
openpyxl==3.0.10
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pyodbc==4.0.34
python-dateutil==2.8.2
python-ldap==3.4.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
